refactor: unit 삭제 시 외래키 제약 조건 문제 해결 및  unit 삭제/수정 api path parameter 방식으로 변경

+++ backend/common/infrastructures/repositories/prAdminUnitRepository.ts
@@ -62,11 +62,20 @@ export class prAdminUnitRepository implements IAdminUnitRepository {
 
   // 단원 삭제
   async delete(id: number): Promise<Unit> {
-    const unit = await this.prisma.unit.delete({
+    const unitToDelete = await this.prisma.unit.findUnique({
       where: { id },
     });
 
-    return this.mapToUnit(unit);
+    if (!unitToDelete) {
+      throw new Error('삭제할 단원을 찾을 수 없습니다.');
+    }
+
+    // Unit 삭제 (Cascade Delete가 관련 데이터 자동 삭제)
+    await this.prisma.unit.delete({
+      where: { id },
+    });
+
+    return this.mapToUnit(unitToDelete);
   }
 
   // 특정 단원의 vidUrl만 조회
